+++ android/anc/src/test/java/org/smartregister/fhircore/anc/app/fakes/FakeModel.kt
@@ -27,6 +27,7 @@ import org.hl7.fhir.r4.model.IntegerType
 import org.hl7.fhir.r4.model.Observation
 import org.hl7.fhir.r4.model.Patient
 import org.hl7.fhir.r4.model.Period
+import org.hl7.fhir.r4.model.Quantity
 import org.hl7.fhir.r4.model.Reference
 import org.hl7.fhir.r4.model.StringType
 
@@ -87,6 +88,12 @@ object FakeModel {
   fun getObservation(testValue: Int = 4) =
     Observation().apply {
       value = IntegerType(testValue)
+      effective = DateTimeType.now()
+    }
+
+  fun getObservationQuantity(testValue: Double = 4.0) =
+    Observation().apply {
+      value = Quantity(testValue)
       effective = DateTimeType.now()
     }
 }