+++ android/quest/src/test/java/org/smartregister/fhircore/quest/ui/patient/details/QuestPatientDetailViewModelTest.kt
@@ -22,6 +22,7 @@ import dagger.hilt.android.testing.HiltAndroidRule
 import dagger.hilt.android.testing.HiltAndroidTest
 import io.mockk.mockk
 import javax.inject.Inject
+import org.hl7.fhir.r4.model.Questionnaire
 import org.hl7.fhir.r4.model.QuestionnaireResponse
 import org.junit.Assert
 import org.junit.Before
@@ -111,5 +112,58 @@ class QuestPatientDetailViewModelTest : RobolectricTest() {
       "67890",
       questPatientDetailViewModel.questionnaireConfigs.value!!.last().identifier
     )
+  }
+
+  @Test
+  fun testFetchResultNonNullNameShouldReturnNameValue() {
+    val result =
+      questPatientDetailViewModel.fetchResultItemLabel(
+        testResult =
+          Pair(
+            QuestionnaireResponse(),
+            Questionnaire().apply {
+              name = "Sample name"
+              title = "Sample title"
+            }
+          )
+      )
+
+    Assert.assertEquals("Sample name", result)
+    Assert.assertNotEquals("Sample title", result)
+  }
+
+  @Test
+  fun testFetchResultNullNameShouldReturnTitleValue() {
+    val result =
+      questPatientDetailViewModel.fetchResultItemLabel(
+        testResult =
+          Pair(
+            QuestionnaireResponse(),
+            Questionnaire().apply {
+              name = null
+              title = "Sample title"
+            }
+          )
+      )
+
+    Assert.assertEquals("Sample title", result)
+    Assert.assertNotEquals("Sample name", result)
+  }
+
+  @Test
+  fun testFetchResultNullNameTitleShouldReturnNull() {
+    val result =
+      questPatientDetailViewModel.fetchResultItemLabel(
+        testResult =
+          Pair(
+            QuestionnaireResponse(),
+            Questionnaire().apply {
+              name = null
+              title = null
+            }
+          )
+      )
+
+    Assert.assertNull(result)
   }
 }